requirements.txt
apache-airflow-providers-postgres==6.5.1
dbt-postgres==1.10.0
great-expectations==1.11.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.45
pandas==2.3.3
numpy==2.1.0
requests==2.32.3
python-dotenv==1.1.1